docs: setup mkdocs for github pages
Sets up mkdocs config and folder structure for building a documentation
site for this repo. Adds a docker-compose environment running the
mkdocs watcher for updating documentation locally.

Adds some basic docs around contributing, and configures themes
and plugins that we'll want.

Adds a github action to deploy the documentation site to github
pages.