build: setup goreleaser (#5)
Adds a version command to the tool, using values from ldflags to
embed it. Updates local build script to pass these in.

Adds a goreleaser config, mostly boilerplate, with a change to
the main location.

Adds a github action to create a new release for each semver tag.

Adds a github action that runs the goreleaser build for each PR,
to verify that the build is working correctly.