build: setup semantic-release (#6)
* build: setup semantic-release

Sets up semantic release to generate the next tag for the repo when
changes are pushed to main. This should only generate the tag,
according to the rules around conventional commits. The release flow
using goreleaser will handle creating the actual release in github
from the new tag..

* build: fix run-name in build workflow

Updates the run name in the ci_build (previously create_pre_release)
workflow to use the PR number; it was just a bug from earlier.

Also updates the name of the workflow to better reflect what it
actually does.