fix: finish check command (#12)
Fills in the logic for the check command. It looks for docker, git,
and docker compose on the system. It will return an error and a
non-zero exit code if any of these tools don't exist; a zero exit
code if they do.

Also adds a tui package for coloured output.

Sets up a service container in the main package for building any
services we need.

Adds a common package where we'll contain the primary "language" of
the application. It will be available to any other package, but it
should use no other package. Provides a simple way to create some
shared structures without cyclical dependencies.