Use tui inside the check command & some code analysis setup (#18)
* refactor: update how the sys check command works

Starts sending output from the handler instead of doing it in the
cobra handler. This allows us to confirm positive outcomes as well,
and it now says which tools are being checked and their respective
results.

* chore: run go fmt on codebase

Runs go fmt across all files to fix them.

Adds a make command and script to run the go fmt command.

* chore: setup go vet

Adds a make command to run go vet across the codebase.

* ci: setup go fmt check

Adds a new github actions workflow for static analysis  that currently
only checks if there are any formatting violations.

* docs: add docs for the new static analysis workflow

Adds an entry to the github actions docs for the static analysis
workflow.