feat: add branches and checkout commands
Add a 'g' group for git commands to the application, within this
there are two new commands: branches, and co.

The branches command shows the branches available in the current
repository, and optionally searches them. The search is very basic
just checking for the search string within the branch name.

The co command uses the same logic as above, you can enter a branch
name that you'd like to checkout and it will search branches for a
match. If there is only one match, it will just check it out. If
there are multiple matches, then it will present an interactive
choice where you can use the TUI to select which branch you actually
want.

To implement the choices TUI, I used the bubbletea package. This
handles the interactivity for the app.

To implement the git commands, it is using the exec package, but i
introduced a service within the hostsys package to abstract this a
bit and avoid some duplication each time we need to call execute a
tool in the system.

While doing this I also introduced a logging package that we can use
to configure log levels throughout the app. As most of this is a
kind of TUI, the errors are presented anyway so the log level by
default is a custom "none" level that won't show anything. This log
level is a persistent flag that can be changed on each command and
will show more errors if needed for debugging.