feat(git): add pull command
Adds a command to pull the current branch that is checked out.

Refactors the executor so that we can more easily handle different
IO requirements. New option to use Host IO so that all stdin, stdout,
and stderr are passed the command. Also an option to pass a get a
buffer for stdout and stderr respectively so the output can easily
be captured.