refactor(git): change how we pull branches
Updates the way we pulled the branch to pass the arguments for origin
and branch name. This is pretty handy to avoid other errors when
pulling branches where we haven't set the upstream name.

Updates how the checkout works so that it returns the name of the
branch that was actually checked out. This was required so that we
can ensure we pull the correct branch after it's checked out.