Add pull option to checkout (#20)
* feat(git): add pull command

Adds a command to pull the current branch that is checked out.

Refactors the executor so that we can more easily handle different
IO requirements. New option to use Host IO so that all stdin, stdout,
and stderr are passed the command. Also an option to pass a get a
buffer for stdout and stderr respectively so the output can easily
be captured.

* feat(git): add a pull option to checkout

Adds a --pull (or -p) option to the 'g co' command. This causes the
branch to be pulled immediately after it's successfully checked out.

Also changes the checkout command to use host IO so we can see
progress and such from git.

* fix(git): add better output if branch is already checked out

Adds a check for the checkout command, so that if the search term
is the same as the current branch, it tells you that instead of
giving what looks like an error message.

* feat(git): add a special case for minus in checkout

Adds a special case to checkout when passing "-". Instead of
searching branches it just passes this straight to the checkout
command to checkout the previous branch.

* refactor(git): change how we pull branches

Updates the way we pulled the branch to pass the arguments for origin
and branch name. This is pretty handy to avoid other errors when
pulling branches where we haven't set the upstream name.

Updates how the checkout works so that it returns the name of the
branch that was actually checked out. This was required so that we
can ensure we pull the correct branch after it's checked out.