refactor(config): improve config management
Refactors how we're managing the config, mainly by moving the single
global variable out of the main package and storing it in the config
manager. Will need a better way to work with it, that we can prevent
modifications without going the config manager, but allow reads.

Currently we'll end up with some cyclical dependencies probably if
we try to avoid knowing anything about the structure of the config
outside of the config package. This should hopefully get into a
better position to support a 'ws ls' command to show information
about the available workspaces.