refactor: update how config works (again)
The config was proving pretty difficult to work with. So I've gone
through and updated how the config worked. There's no config manager
any more, we just have a Config struct, that internally has 2 copies
of the config. One copy is kep in sync with whatever is on disk,
and is also used when making persistent changes. This is to account
for the fact that certain things (e.g. log level) can be changed
temporarily in the CLI and should not be reflected in the persisted
config.

Then there is a runtime config which has these temporary overrides
applied, and should typically be used for any read operations on
the configs.

Also introduced a slice package which we can use for udpating the
slices of projects/workspaces with much less effort/repeition of
logic. This makes the mutations feel much simpler than previously.

The Config can now be better used as a simple interface we can
pass around, and any changes are automatically saved from within
the mutators rather than the consumer.

Added a bunch of tests to make sure this all works how I expect.

wg