feat: start creating up command
Adds the up command to start the workspace. Determines the context
of which projects to start based on a couple of conditions. If inside
a project repository, only that project will be started. If inside
a workspace repository, all projects for that workspace will be
started. If the -p option is supplied it overrides the project
repository rule, and will start the specified project only (if the
project is part of the workspace chosen as well). If the workspace
option is not supplied, we'll default to the current workspace from
the config. If no options are supplied and we're not in a project
or workspace repo, then we'll start up all services within the
current workspace (as defined in the config).

Determines the startup order by utilising a DAG; a new package was
added to build the DAG and sort the projects according to what they
require. There is a little ambiguity in the exact order, as some
projects could be started at the same time. In this case the order
doesn't really matter, and will be done alphabetically, due to how
the DAG package sorts the leaves/roots of the graph.

At this stage the up command will only display the projects configs
in the order it would start them up. It's enough to start building
upon in later changes.

This also included a fairly heavy refactor of how we're loading
configurations for projects. This was necessary to avoid the places
that need then configs having to load and stitch together
multiple different structs into a coherent context.