feat(overlays): add overlay generation for networks
Adds the overlays feature. This feature is responsible for inspecting
the compose files in each project, and generating an extra compose
file that can be included to add the relevant network configuration
to each service in each project. This allows multiple projects to
access eachother across the network.

As part of this, also added an extra debug command that exposes
a sub command that shows the final compose config used once
everything has been merged together. Very useful for checking that
the overlays are having the correct affect.