feat(overlays): add aliases and labels with network overlay
Updates the network overlay to also add aliases to each service.
The alias can be configured via a template that allows you to use
the service name, project name and workspace name to build a URL.

Also adds labels to the services that have been affected by overlays,
mostly so if we inspect the containers, we can see that an overlay
modified the service.