fix: fix vulnerabilities in dockerfile
Fixes some vulnerabilities in the dockerfile for the local docs
server. Mostly minor (considering this is only used for local dev)
but simple enough to fix rather than avoid scanning.

Needed to add a user, not use root (this one is pretty obvious, and
would be an issue for anything production level), needed to pin to
a more specific version than latest, and finally needed to add a
healthcheck.