feat(overlays): add a tls injection overlay
Adds a TLS injection overlay, that is enabled on per-service basis
by adding a label in the docker compose file. When enabled this will
inject the all the generated TLS certificates into the services
container. The path at which they're mounted is controlled by the
labels value.

One quirk, due to how the tls generation was setup it will inject
ALL certificates for all workspaces into the container. This is
probably not an issue generally, but something to consider tidying
up later.