test: change cover mode for go test
Fixes an issue in CI where a file was moved, and this broke the
coverage action. It seems to only be an issue when using the atomic
cover mode and not for the default 'set'; we dont' actually need
atomic, it just came from the example for the test-coverage action.