fix: run docker compose commands from correct directory
Fixes an issue when running a docker compose command from outside
the project directory, where the overlays and env files are not
properly included. This was most notable when running up or down
for an entire project as the overlays didn't render properly and
gave a warning. The fix is to actually change the working directory
to each project when running any commands for it; it will not change
the actual working dir for the user in the terminal.