feat: support exec or ext in non-running containers
Adds support for running commands inside a service that isn't actually
running. This is useful for helper scripts like migrations that only
need to run occasionally, and at manual request. This supports this
by checking if the service is running and using the run command
instead of the exec command if it isn't.

Also addressed an issue with the networking when a profile isn't
supplied. When generating the overlays any services that weren't
enabled due to a profile not being supplied, didn't get the relevant
networking config. This fixes that by enabling all profiles when
generating the overlays.